Normalize all cutscene functions
