+++ config/splat.us.stcen.yaml
@@ -41,7 +41,7 @@ segments:
       - [0x394, .data, e_init]
       - [0x4D4, .data, st_debug]
       - [0x554, .data, e_breakable]
-      - [0x5D8, .data, holyglassescutscene]
+      - [0x5D8, .data, cutscene]
       - [0x68C, .data, F890]
       - [0x7B8, .data, st_update]
       - [0x7E4, .data, collision]
@@ -68,7 +68,7 @@ segments:
       - [0x81EC, .data, tile_data] # tile definitions
       - [0xC60C, .data, sprites]
       - [0xD414, .rodata, e_init]
-      - [0xD42C, .rodata, holyglassescutscene]
+      - [0xD42C, .rodata, cutscene]
       - [0xD4B0, .rodata, F890]
       - [0xD4D8, .rodata, e_red_door] # EntityRedDoor
       - [0xD540, .rodata, e_collect] # EntityPrizeDrop
@@ -80,7 +80,7 @@ segments:
       - [0xD600, c, st_debug]
       - [0xD89C, c, e_breakable]
       - [0xDB18, c]
-      - [0xDF0C, c, holyglassescutscene]
+      - [0xDF0C, c, cutscene]
       - [0xF890, c]
       - [0x10E4C, c, st_update]
       - [0x11280, c, collision]
@@ -97,6 +97,6 @@ segments:
       - [0x1BED4, c, prim_helpers]
       - [0x1C704, .bss, header]
       - [0x1C764, .bss, create_entity]
-      - [0x1C774, .bss, holyglassescutscene]
+      - [0x1C774, .bss, cutscene]
       - [0x1C7C0, .bss, e_collect]
   - [0x1D46C]
