+++ config/splat.us.stno3.yaml
@@ -76,7 +76,7 @@ segments:
       - [0x37414, .rodata, 3C4EC]
       - [0x3742C, .rodata, 3E134] #EntityFlyingOwlAndLeaves
       - [0x37440, rodata]
-      - [0x37458, .rodata, deathcutscene] # EntityDeathCutscene
+      - [0x37458, .rodata, cutscene] # NO3_CutsceneExec
       - [0x374DC, .rodata, 3FF00] # EntityRoomTransition2
       - [0x374F4, .rodata, 3FF00] # EntityDeathStolenItems
       - [0x37514, .rodata, 3FF00] # EntityDeath
@@ -102,7 +102,7 @@ segments:
       - [0x377D4, c]
       - [0x3C4EC, c]
       - [0x3E134, c]
-      - [0x3E544, c, deathcutscene]
+      - [0x3E544, c, cutscene]
       - [0x3FF00, c]
       - [0x4184C, c, st_update]
       - [0x41C80, c, collision]
