+++ config/splat.us.stst0.yaml
@@ -40,7 +40,7 @@ segments:
       - [0x4BC, data]
       - [0x580, data]
       - [0x770, data]
-      - [0x828, .data, draculacutscene] # "Richter, Dracula"
+      - [0x828, .data, cutscene] # "Richter, Dracula"
       - [0x830, data]
       - [0x990, data]
       - [0xAB0, data]
@@ -66,7 +66,7 @@ segments:
       - [0x1E6D0, .data, tile_data] # tile definitions
       - [0x226E0, .data, sprites]
       - [0x27984, .rodata, 27D64]
-      - [0x279E8, .rodata, draculacutscene] # EntityDraculaCutscene
+      - [0x279E8, .rodata, cutscene] # ST0_CutsceneExec
       - [0x27A74, .rodata, 2A218] # EntityCutscene
       - [0x27AB4, .rodata, 2A8DC] # EntitySecretButton
       - [0x27ACC, .rodata, 2A8DC] # EntitySecretStairs
@@ -88,7 +88,7 @@ segments:
       - [0x27D3C, .rodata, prim_helpers]
       - [0x27D44, .rodata, 3D8F0]
       - [0x27D64, c]
-      - [0x28984, c, draculacutscene]
+      - [0x28984, c, cutscene]
       - [0x2A218, c]
       - [0x2A8DC, c]
       - [0x2B0C8, c]
