+++ config/symbols.us.stno3.txt
@@ -64,7 +64,7 @@ EntityBackgroundTrees = 0x801BDEB0;
 EntityBackgroundCastleWall = 0x801BE080;
 EntityFlyingOwlAndLeaves = 0x801BE134;
 EntityFallingLeaf = 0x801BE444;
-EntityDeathCutscene = 0x801BEEB0;
+NO3_CutsceneExec = 0x801BEEB0;
 EntityRoomTransition2 = 0x801BFF00;
 EntityDeathStolenItem = 0x801C033C;
 EntityDeath = 0x801C083C;
