+++ config/symbols.us.stsel.txt
@@ -1,6 +1,6 @@
 SEL_g_EntityGfxs = 0x80180760;
 g_Streams = 0x8018B2EC;
-func_801B69F8 = 0x801B69F8;
+SEL_CutsceneExec = 0x801B69F8;
 StreamDiskIsReady = 0x801B9744;
 StreamNextVlc = 0x801B988C;
 func_801BAB70 = 0x801BAB70;
