+++ config/symbols.us.stst0.txt
@@ -19,7 +19,7 @@ D_80197F80 = 0x80197F80;
 D_8019A40C = 0x8019A40C;
 c_HeartPrizes = 0x801A7C84;
 EntityLockCamera = 0x801A7EB0;
-EntityDraculaCutscene = 0x801A9210;
+ST0_CutsceneExec = 0x801A9210;
 EntityCutscene = 0x801AA290;
 EntitySecretButton = 0x801AA8DC;
 EntitySecretStairsCeiling = 0x801AABA4;
@@ -57,3 +57,4 @@ EntityClockTower3D = 0x801BB2AC;
 EntityCutscenePhotograph = 0x801BB764;
 EntityCutscenePhotographFire = 0x801BC4D0;
 EntityBackgroundVortex = 0x801BE1B4;
+skip_cutscene = 0x801C24C8;
