+++ src/st/cen/cutscene.c
@@ -1,73 +1,120 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-#include "no3.h"
+#include "cen.h"
+#include <cutscene.h>
 
-// similar but not an exact duplicate
-void CutsceneUnk1(void) {
-    g_Dialogue.nextLineX = 0x182; // Note that these two lines are "= 2"
-    g_Dialogue.nextCharX = 0x182; // for all other cutscenes (so far)!
-    g_Dialogue.nextCharY = 0;
-    g_Dialogue.unk12 = 0;
-    g_Dialogue.nextCharTimer = 0;
-    g_Dialogue.unk17 = 8;
-    g_Dialogue.nextLineY = g_Dialogue.startY + 0x14;
-}
+static u8 __unused[0xC00];
+static s32 skip_cutscene;
+static Dialogue g_Dialogue;
+static u32 __unused0[26];
+u32 g_CutsceneFlags;
+u32 D_8019D428;
+
+// Bizarre variable - u8 here, but u16 in EntityHeartDrop
+// Possible that since EntityHeartDrop is not used in CEN, that data
+// is reused here for the holy glasses cutscene? Very unusual.
+u8 g_HeartDropArray[] = {0, 64, 0, 0};
+static u8 D_801805DC[] = {0, 0, 0, 0};
+
+// clut
+static u16 D_801805E0[] = {
+    LAYER_CLUT_ALT | 0,
+    LAYER_CLUT_ALT | 8,
+    LAYER_CLUT_ALT | 16,
+    LAYER_CLUT_ALT | 24,
+};
+static u16 D_801805E8[] = {0, 32};
+
+static u16 D_801805EC[] = {
+    0x01A1,
+    0x01A1,
+    0x01A1,
+    0x01A1,
+};
+
+// x-coords
+static s16 D_801805F4[] = {
+    0x08, 0x13, 0x11, 0x31, 0x4F, 0x26, 0x36, 0x1D, 0x1B, 0x33, 0x2C, 0x21,
+    0x19, 0x0A, 0x33, 0x1F, 0x48, 0x2F, 0x13, 0x19, 0x4D, 0x4B, 0x17, 0x1D,
+    0x12, 0x02, 0x1B, 0x2A, 0x50, 0x45, 0x32, 0x0D, 0x2A, 0x4D, 0x06, 0x27,
+    0x07, 0x48, 0x2F, 0x1B, 0x36, 0x22, 0x39, 0x14, 0x39, 0x1D, 0x0A, 0x35,
+    0x10, 0x1B, 0x3D, 0x17, 0x2E, 0x0B, 0x49, 0x42, 0x3D, 0x2A, 0x01, 0x0C,
+    0x1B, 0x34, 0x41, 0x35, 0x08, 0x0E, 0x4D, 0x11, 0x34, 0x41, 0x29, 0x48,
+};
+
+static const char* g_ActorNames[] = {
+    _S("Alucard"),
+    _S("Maria"),
+};
+
+static const char _pad[4] = "";
+
+#include "../cutscene_unk1.h"
 
 #include "../set_cutscene_script.h"
 
-void CutsceneUnk3(s16 yOffset) {
-    RECT rect;
+#include "../cutscene_unk3.h"
 
-    rect.x = 384;
-    rect.y = (yOffset * 12) + 256;
-    rect.w = 64;
-    rect.h = 12;
-    ClearImage(&rect, 0, 0, 0);
-}
-
-void CutsceneUnk4(void) {
-    Primitive* prim;
-
-    CutsceneUnk3(g_Dialogue.nextCharY);
-    prim = g_Dialogue.prim[g_Dialogue.nextCharY];
-    prim->tpage = 0x16;
-    prim->clut = g_Dialogue.clutIndex;
-    prim->y0 = g_Dialogue.nextLineY;
-    prim->u0 = 0;
-    prim->x0 = g_Dialogue.startX;
-    prim->x0 = prim->x0 + 4;
-    prim->v0 = g_Dialogue.nextCharY * 0xC;
-    prim->u1 = 0xC0;
-    prim->v1 = 0xC;
-    prim->priority = 0x1FF;
-    prim->drawMode = DRAW_DEFAULT;
-}
+#include "../cutscene_unk4.h"
 
 #include "../cutscene_avatar.h"
 
 #include "../cutscene_unk6.h"
 
-#include "../cutscene.h"
+static void CutsceneRun(void) {
+    Entity* entity;
+    u16 startTimer;
 
-void func_801BECCC(Entity* self) {
-    if (g_pads[0].tapped == PAD_START) {
-        D_801D7D20 = 1;
-        g_api.FreePrimitives(self->primIndex);
-        self->flags ^= FLAG_HAS_PRIMS;
-        if (g_Dialogue.primIndex[1] != -1) {
-            g_api.FreePrimitives(g_Dialogue.primIndex[1]);
+    g_Dialogue.timer++;
+    // protect from overflows
+    if (g_Dialogue.timer >= 0xFFFF) {
+        g_Dialogue.unk3C = 0;
+        return;
+    }
+    while (true) {
+        // Start the dialogue script only if the start timer has passed
+        startTimer = *g_Dialogue.unk40++ << 8;
+        startTimer |= *g_Dialogue.unk40++;
+        if (g_Dialogue.timer < startTimer) {
+            // Re-evaluate the condition at the next frame
+            g_Dialogue.unk40 -= 2;
+            return;
         }
-        if (g_Dialogue.primIndex[0] != -1) {
-            g_api.FreePrimitives(g_Dialogue.primIndex[0]);
+        switch (*g_Dialogue.unk40++) {
+        case 0:
+            entity =
+                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
+            DestroyEntity(entity);
+            entity->entityId = *g_Dialogue.unk40++;
+            entity->pfnUpdate = PfnEntityUpdates[entity->entityId - 1];
+            entity->posX.i.hi = *g_Dialogue.unk40++ * 0x10;
+            entity->posX.i.hi |= *g_Dialogue.unk40++;
+            entity->posY.i.hi = *g_Dialogue.unk40++ * 0x10;
+            entity->posY.i.hi |= *g_Dialogue.unk40++;
+            entity->posX.i.hi -= g_Tilemap.scrollX.i.hi;
+            entity->posY.i.hi -= g_Tilemap.scrollY.i.hi;
+            break;
+        case 1:
+            entity =
+                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
+            DestroyEntity(entity);
+            break;
+        case 2:
+            if (!((g_CutsceneFlags >> *g_Dialogue.unk40) & 1)) {
+                g_Dialogue.unk40--;
+                return;
+            }
+            g_CutsceneFlags &= ~(1 << *g_Dialogue.unk40++);
+            break;
+        case 3:
+            g_CutsceneFlags |= 1 << *g_Dialogue.unk40++;
+            break;
         }
-        g_api.PlaySfx(SET_STOP_MUSIC);
-        self->step = 1;
-        self->step_s = 0;
     }
 }
 
 #include "../cutscene_scale_avatar.h"
 
-void EntityDeathCutscene(Entity* self) {
+void CEN_CutsceneExec(Entity* self) {
     RECT rect;
     Primitive* prim;
     s32 primIndex;
@@ -77,33 +124,41 @@ void EntityDeathCutscene(Entity* self) {
     s16 vCoord;
     u16 nextChar;
     s32 nextChar2;
-    s32 bit_shifty;
+    u_long ptr;
 
-    if (self->step != 0) {
-        if ((D_801D7DD4 != 0) && (D_801D7D20 == 0) &&
-            ((g_IsTimeAttackUnlocked) || (g_Settings.D_8003CB04 & 4))) {
-            func_801BECCC(self);
+    if (self->step) {
+        if (D_8019D428 && !skip_cutscene &&
+            ((g_Settings.D_8003CB04 & 0x800) || g_IsTimeAttackUnlocked) &&
+            (g_pads[0].tapped & PAD_START)) {
+            skip_cutscene = 1;
+            g_api.FreePrimitives(self->primIndex);
+            self->flags ^= FLAG_HAS_PRIMS;
+            if (g_Dialogue.primIndex[1] != -1) {
+                g_api.FreePrimitives(g_Dialogue.primIndex[1]);
+            }
+            if (g_Dialogue.primIndex[0] != -1) {
+                g_api.FreePrimitives(g_Dialogue.primIndex[0]);
+            }
+            g_api.PlaySfx(SET_STOP_MUSIC);
+            self->step = 1;
+            self->step_s = 0;
         }
-        if ((self->step != 0) && (D_801D7D60 != 0)) {
+        if (self->step && (g_Dialogue.unk3C != 0)) {
             CutsceneRun();
         }
     }
     switch (self->step) {
     case 0:
-        CreateEntityFromCurrentEntity(E_BG_LIGHTNING, &g_Entities[192]);
-        if ((g_CastleFlags[CASTLE_FLAG_53]) ||
-            (g_PlayableCharacter != PLAYER_ALUCARD)) {
+        if (g_CastleFlags[HG_CS_DONE] != 0) {
             DestroyEntity(self);
-            g_Entities[192].params = 0;
             return;
         }
-        g_Entities[192].params = 0x100;
-        if (SetCutsceneScript(D_80184CE0)) {
+        if (SetCutsceneScript(g_CutsceneScript)) {
             self->flags |= FLAG_HAS_PRIMS | FLAG_UNK_2000;
-            g_CutsceneFlags = 0;
-            D_801D7DD4 = 0;
-            D_801D7D20 = 0;
             D_8003C704 = 1;
+            g_CutsceneFlags = 0;
+            D_8019D428 = 0;
+            skip_cutscene = 0;
             self->primIndex = g_Dialogue.primIndex[2];
             self->step++;
         }
@@ -111,17 +166,17 @@ void EntityDeathCutscene(Entity* self) {
     case 1:
         // this is a huge While-loop! Don't miss it!
         while (1) {
-            if ((g_Dialogue.nextCharTimer != 0) && (D_801D7D20 == 0)) {
+            if ((g_Dialogue.nextCharTimer != 0) && !skip_cutscene) {
                 g_Dialogue.nextCharTimer--;
                 return;
             }
             nextChar = *g_Dialogue.nextCharDialogue++;
             switch (nextChar) {
-            case 0:
+            case CSOP_END_CUTSCENE:
                 self->step = 7;
                 return;
-            case 1:
-                if (D_801D7D20 != 0) {
+            case CSOP_LINE_BREAK:
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharX = g_Dialogue.nextLineX;
@@ -147,17 +202,17 @@ void EntityDeathCutscene(Entity* self) {
                 self->step_s = 0;
                 self->step++;
                 return;
-            case 2:
+            case CSOP_SET_SPEED:
                 g_Dialogue.unk17 = *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 3:
+            case CSOP_SET_WAIT:
                 g_Dialogue.nextCharTimer = *g_Dialogue.nextCharDialogue++;
-                if (D_801D7D20 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 return;
-            case 4:
-                if (D_801D7D20 != 0) {
+            case CSOP_SCRIPT_UNKNOWN_4:
+                if (skip_cutscene) {
                     continue;
                 }
                 prim = g_Dialogue.prim[0];
@@ -166,8 +221,8 @@ void EntityDeathCutscene(Entity* self) {
                     prim = prim->next;
                 }
                 return;
-            case 5:
-                if (D_801D7D20 != 0) {
+            case CSOP_SET_PORTRAIT:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 2;
                     continue;
                 }
@@ -175,10 +230,10 @@ void EntityDeathCutscene(Entity* self) {
                 i = *g_Dialogue.nextCharDialogue++;
                 nextChar2 = *g_Dialogue.nextCharDialogue++;
                 prim = g_Dialogue.prim[5];
-                uCoord = D_80181A28[nextChar2 & 1];
-                vCoord = D_80181A2C[nextChar2 & 1];
-                prim->clut = D_80181A30[i];
-                prim->tpage = 0x94; // Weird, this is 0x90 in other cutscenes
+                uCoord = g_HeartDropArray[nextChar2 & 1];
+                vCoord = D_801805DC[nextChar2 & 1];
+                prim->clut = D_801805E0[i];
+                prim->tpage = 0x90;
                 if (nextChar2 & 0x80) {
                     prim->u0 = prim->u2 = uCoord + 0x2F;
                     prim->u1 = prim->u3 = uCoord;
@@ -192,7 +247,7 @@ void EntityDeathCutscene(Entity* self) {
                     g_Dialogue.startX - 0x1E;
                 prim->y0 = prim->y1 = prim->y2 = prim->y3 =
                     g_Dialogue.startY + 0x24;
-                g_Dialogue.clutIndex = D_80181A38[i];
+                g_Dialogue.clutIndex = D_801805EC[i];
                 CutsceneUnk1();
                 CutsceneUnk4();
                 prim->priority = 0x1FE;
@@ -201,8 +256,8 @@ void EntityDeathCutscene(Entity* self) {
                 g_Dialogue.portraitAnimTimer = 6;
                 self->step = 3;
                 return;
-            case 6:
-                if (D_801D7D20 != 0) {
+            case CSOP_NEXT_DIALOG:
+                if (skip_cutscene) {
                     continue;
                 }
                 for (prim = g_Dialogue.prim[0], i = 0; i < 5; i++) {
@@ -214,8 +269,8 @@ void EntityDeathCutscene(Entity* self) {
                 g_Dialogue.portraitAnimTimer = 6;
                 self->step = 4;
                 return;
-            case 7:
-                if (D_801D7D20 != 0) {
+            case CSOP_SET_POS:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -233,16 +288,15 @@ void EntityDeathCutscene(Entity* self) {
                 self->step = 5;
                 self->step_s = 0;
                 return;
-
-            case 8:
-                if (D_801D7D20 != 0) {
+            case CSOP_CLOSE_DIALOG:
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.portraitAnimTimer = 0x18;
                 self->step = 6;
                 return;
-            case 9:
-                if (D_801D7D20 != 0) {
+            case CSOP_PLAY_SOUND:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -252,8 +306,8 @@ void EntityDeathCutscene(Entity* self) {
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            case 10:
-                if (D_801D7D20 != 0) {
+            case CSOP_WAIT_FOR_SOUND:
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != false) {
@@ -261,8 +315,8 @@ void EntityDeathCutscene(Entity* self) {
                 }
                 *g_Dialogue.nextCharDialogue--;
                 return;
-            case 11:
-                if (D_801D7D20 != 0) {
+            case CSOP_UNK_11:
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != true) {
@@ -270,51 +324,49 @@ void EntityDeathCutscene(Entity* self) {
                 }
                 *g_Dialogue.nextCharDialogue--;
                 return;
-            case 12:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                CutsceneUnk6((u8*)bit_shifty);
+            case CSOP_SCRIPT_UNKNOWN_12:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                CutsceneUnk6((u8*)ptr);
                 continue;
-            case 13:
+            case CSOP_SCRIPT_UNKNOWN_13:
                 continue;
-            case 14:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty += (u16)0x100000;
-                g_Dialogue.nextCharDialogue += *(u16*)bit_shifty << 2;
+            case CSOP_SCRIPT_UNKNOWN_14:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr += (u16)0x100000;
+                g_Dialogue.nextCharDialogue += *(u16*)ptr << 2;
 
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
-
-            case 15:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+            case CSOP_SCRIPT_UNKNOWN_15:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
-
-            case 16:
+            case CSOP_WAIT_FOR_FLAG:
                 if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
                     g_Dialogue.nextCharDialogue--;
                     return;
@@ -322,46 +374,46 @@ void EntityDeathCutscene(Entity* self) {
                 g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue);
                 *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 17:
+            case CSOP_SET_FLAG:
                 g_CutsceneFlags |= 1 << *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 18:
+            case CSOP_SCRIPT_UNKNOWN_18:
                 g_Dialogue.unk3C = 0;
                 continue;
-            case 19:
-                if (D_801D7D20 != 0) {
+            case CSOP_LOAD_PORTRAIT:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 5;
                 } else {
-                    bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty += 0x100000;
+                    ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr += 0x100000;
                     nextChar2 = g_Dialogue.nextCharDialogue++[0];
-                    LoadTPage((u32*)bit_shifty, 1, 0, D_80181A34[nextChar2],
-                              0x100, 0x30, 0x48);
+                    LoadTPage((u_long*)ptr, 1, 0, D_801805E8[nextChar2], 0x100,
+                              48, 72);
                 }
                 continue;
-            case 20:
+            case CSOP_SCRIPT_UNKNOWN_20:
                 nextChar = *g_Dialogue.nextCharDialogue++;
                 nextChar <<= 4;
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            case 21:
+            case CSOP_SCRIPT_UNKNOWN_21:
                 g_CutsceneFlags = 0;
-                D_801D7D20 = 0;
-                D_801D7DD4 = 0;
+                skip_cutscene = 0;
+                D_8019D428 = 0;
                 continue;
-            case 22:
+            case CSOP_SCRIPT_UNKNOWN_22:
                 g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue++);
                 continue;
-            case 23:
+            case CSOP_SCRIPT_UNKNOWN_23:
                 return;
-            case 24:
+            case CSOP_SCRIPT_UNKNOWN_24:
                 if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
                     *g_Dialogue.nextCharDialogue--;
                     return;
@@ -369,7 +421,7 @@ void EntityDeathCutscene(Entity* self) {
                 *g_Dialogue.nextCharDialogue++;
                 continue;
             default:
-                if (D_801D7D20 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharTimer = g_Dialogue.unk17;
@@ -384,8 +436,7 @@ void EntityDeathCutscene(Entity* self) {
         rect.y = ((u32)(nextChar & 0xF0) >> 1) + 0xF0;
         rect.w = 2;
         rect.h = 8;
-        // Other cutscenes have + 0x180 here
-        vCoord = (g_Dialogue.nextCharY * 0xC) + 0x100;
+        vCoord = (g_Dialogue.nextCharY * 0xC) + 0x180;
         MoveImage(&rect, g_Dialogue.nextCharX, vCoord);
         g_Dialogue.nextCharX += 2;
         break;
@@ -429,7 +480,7 @@ void EntityDeathCutscene(Entity* self) {
     case 5:
         switch (self->step_s) {
         case 0:
-            D_801D7DD4 = 1;
+            D_8019D428 = 1;
             primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 0x48);
             if (primIndex == -1) {
                 DestroyEntity(self);
@@ -444,7 +495,7 @@ void EntityDeathCutscene(Entity* self) {
                 prim->y0 = prim->y1 = g_Dialogue.startY + j;
                 prim->priority = 0x1FE;
                 prim->drawMode = DRAW_DEFAULT;
-                prim->x2 = D_80181A3C[j];
+                prim->x2 = D_801805F4[j];
                 prim->x3 = 0xF70;
 
                 j++;
@@ -529,12 +580,12 @@ void EntityDeathCutscene(Entity* self) {
         break;
 
     case 7:
-        g_CastleFlags[CASTLE_FLAG_53] = 1;
-        g_api.TimeAttackController(
-            TIMEATTACK_EVENT_MEET_DEATH, TIMEATTACK_SET_RECORD);
-        D_8003C704 = 0;
-        g_Settings.D_8003CB04 |= 4;
         DestroyEntity(self);
+        g_CastleFlags[HG_CS_DONE] = 1;
+        g_api.TimeAttackController(
+            TIMEATTACK_EVENT_GET_HOLYGLASSES, TIMEATTACK_SET_RECORD);
+        D_8003C704 = 0;
+        g_Settings.D_8003CB04 |= 0x800;
         break;
     }
 }
