+++ src/st/cutscene_unk1.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-void CutsceneUnk1(void) {
+static void CutsceneUnk1(void) {
     g_Dialogue.nextLineX = 2;
     g_Dialogue.nextCharX = 2;
     g_Dialogue.nextCharY = 0;
