+++ src/st/cutscene_unk6.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-void CutsceneUnk6(s32 arg0) {
+static void CutsceneUnk6(u8* arg0) {
     g_Dialogue.unk40 = arg0 + 0x100000;
     g_Dialogue.timer = 0;
     g_Dialogue.unk3C = 1;
