+++ src/st/dre/bss.c
@@ -2,7 +2,7 @@
 #include <common.h>
 
 STATIC_PAD_BSS(3072);
-u32 D_801A3ED4;
+u32 skip_cutscene;
 u32 g_Dialogue;
 u16 D_801A3EDC;
 u16 D_801A3EDE;
