+++ src/st/dre/cutscene.c
@@ -1,89 +1,98 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-/*
- * Overlay: ST0
- * Description: Cutscene of Richter talking to Dracula
- */
-
-#include "st0.h"
+#include "dre.h"
+#include <cutscene.h>
 
 #include "../cutscene_unk1.h"
 
-// not an exact duplicate
-u8 SetCutsceneScript(const char* textDialogue) {
-    Primitive* prim;
-    s16 firstPrimIndex;
-
-    firstPrimIndex = g_api.AllocPrimitives(PRIM_SPRT, 8);
-    g_Dialogue.primIndex[2] = firstPrimIndex;
-    if (firstPrimIndex == -1) {
-        g_Dialogue.primIndex[2] = 0;
-        return 0;
-    }
-    g_Dialogue.nextCharDialogue = textDialogue;
-    g_Dialogue.unk3C = 0;
-    g_Dialogue.primIndex[1] = -1;
-    g_Dialogue.primIndex[0] = -1;
-    CutsceneUnk1();
-
-    if (prim && prim) { // !FAKE
-    }
-
-    prim = g_Dialogue.prim[0] = &g_PrimBuf[g_Dialogue.primIndex[2]];
-
-    prim->drawMode = DRAW_HIDE;
-    prim = g_Dialogue.prim[1] = prim->next;
-
-    prim->drawMode = DRAW_HIDE;
-    prim = g_Dialogue.prim[2] = prim->next;
-
-    prim->drawMode = DRAW_HIDE;
-    prim = g_Dialogue.prim[3] = prim->next;
-
-    prim->drawMode = DRAW_HIDE;
-    prim = g_Dialogue.prim[4] = prim->next;
-
-    prim->drawMode = DRAW_HIDE;
-    prim = g_Dialogue.prim[5] = prim->next;
-
-    prim->type = 4;
-    prim->drawMode = DRAW_HIDE;
-
-    prim = prim->next;
-    prim->type = 3;
-    prim->r0 = prim->r1 = prim->r2 = prim->r3 = 0xFF;
-    prim->g0 = prim->g1 = prim->g2 = prim->g3 = 0;
-    prim->b0 = prim->b1 = prim->b2 = prim->b3 = 0;
-    prim->x0 = prim->x2 = 4;
-    prim->x1 = prim->x3 = 0xF8;
-    prim->priority = 0x1FD;
-    prim->drawMode = DRAW_HIDE;
-
-    prim = prim->next;
-    prim->type = 1;
-    prim->x0 = 3;
-    prim->y0 = 0x2F;
-    prim->v0 = 0x4A;
-    prim->r0 = prim->g0 = prim->b0 = 0xFF;
-    prim->priority = 0x1FC;
-    prim->drawMode = DRAW_HIDE;
-    return 1;
-}
+#include "../set_cutscene_script.h"
 
 #include "../cutscene_unk3.h"
 
 #include "../cutscene_unk4.h"
 
-const char* g_ActorNames[] = {_S("Richter"), _S("Dracula")};
+// This array is in Data, but the strings within it are rodata. That rodata
+// lives between the jump tables in this file, so we have to import the rodata,
+// and also the data.
+static const char* g_ActorNames[] = {_S("Alucard"), _S("Lisa"), _S("Succubus")};
 
 #include "../cutscene_avatar.h"
 
 #include "../cutscene_unk6.h"
 
-#include "../cutscene.h"
+void DRE_CutsceneExec(void) {
+    Entity* entity;
+    u16 startTimer;
+
+    g_Dialogue.timer++;
+    // protect from overflows
+    if (g_Dialogue.timer >= 0xFFFF) {
+        g_Dialogue.unk3C = 0;
+        return;
+    }
+    while (true) {
+        // Start the dialogue script only if the start timer has passed
+        startTimer = *g_Dialogue.unk40++ << 8;
+        startTimer |= *g_Dialogue.unk40++;
+        if (g_Dialogue.timer < startTimer) {
+            // Re-evaluate the condition at the next frame
+            g_Dialogue.unk40 -= 2;
+            return;
+        }
+        switch (*g_Dialogue.unk40++) {
+        case 0:
+            entity =
+                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
+            DestroyEntity(entity);
+            entity->entityId = *g_Dialogue.unk40++;
+            entity->pfnUpdate = PfnEntityUpdates[entity->entityId - 1];
+            entity->posX.i.hi = *g_Dialogue.unk40++ * 0x10;
+            entity->posX.i.hi |= *g_Dialogue.unk40++;
+            entity->posY.i.hi = *g_Dialogue.unk40++ * 0x10;
+            entity->posY.i.hi |= *g_Dialogue.unk40++;
+            break;
+        case 1:
+            entity =
+                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
+            DestroyEntity(entity);
+            break;
+        case 2:
+            if (!((g_CutsceneFlags >> *g_Dialogue.unk40) & 1)) {
+                g_Dialogue.unk40 -= 3;
+                g_Dialogue.timer--;
+                return;
+            }
+            g_CutsceneFlags &= ~(1 << *g_Dialogue.unk40++);
+            break;
+        case 3:
+            g_CutsceneFlags |= 1 << *g_Dialogue.unk40++;
+            break;
+        }
+    }
+}
+
+void func_80194F14(Entity* self) {
+    if (g_pads[0].tapped == PAD_START) {
+        skip_cutscene = 1;
+        g_api.FreePrimitives(self->primIndex);
+        self->flags ^= FLAG_HAS_PRIMS;
+        if (g_Dialogue.primIndex[1] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[1]);
+        }
+        if (g_Dialogue.primIndex[0] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[0]);
+        }
+        g_api.PlaySfx(SET_STOP_MUSIC);
+        self->step = 1;
+        self->step_s = 0;
+    }
+}
 
 #include "../cutscene_scale_avatar.h"
 
-void EntityDraculaCutscene(Entity* self) {
+// dialogue with mother opens as alucard walks right ID 20
+// Same pattern as a lot of other cutscenes, main differences
+// are related to the self->params if-else statements.
+void EntitySuccubusCutscene(Entity* self) {
     RECT rect;
     Primitive* prim;
     s32 primIndex;
@@ -93,60 +102,70 @@ void EntityDraculaCutscene(Entity* self) {
     s16 vCoord;
     u16 nextChar;
     s32 nextChar2;
-    s32 bit_shifty;
+    u_long ptr;
 
-    if ((D_801C2580 != 0) && (D_801C24C8 == 0) &&
-        ((g_Settings.D_8003CB04 & 1) || (g_IsTimeAttackUnlocked != 0)) &&
-        (g_pads[0].tapped == PAD_START)) {
-        D_801C24C8 = 1;
-        g_api.FreePrimitives(self->primIndex);
-        self->flags &= ~FLAG_HAS_PRIMS;
-        if (g_Dialogue.primIndex[1] != -1) {
-            g_api.FreePrimitives(g_Dialogue.primIndex[1]);
+    if (self->step != 0) {
+        if ((D_801A3F88 != 0) && !skip_cutscene) {
+            if (g_IsTimeAttackUnlocked) {
+                func_80194F14(self);
+            } else {
+                if (self->params) {
+                    if (g_Settings.D_8003CB04 & 0x2000) {
+                        func_80194F14(self);
+                    }
+                } else if (g_Settings.D_8003CB04 & 0x1000) {
+                    func_80194F14(self);
+                }
+            }
         }
-        if (g_Dialogue.primIndex[0] != -1) {
-            g_api.FreePrimitives(g_Dialogue.primIndex[0]);
+        if ((self->step != 0) && (g_Dialogue.unk3C != 0)) {
+            DRE_CutsceneExec();
         }
-        g_api.PlaySfx(0x80);
-        self->step = 1;
-        self->step_s = 0;
     }
-    if ((self->step) && (g_Dialogue.unk3C != 0)) {
-        CutsceneRun();
-    }
-
     switch (self->step) {
     case 0:
-        if (SetCutsceneScript(D_801829D8)) {
-            self->flags |= FLAG_HAS_PRIMS;
+        if ((g_CastleFlags[SUCC_CS_DONE]) ||
+            (g_PlayableCharacter != PLAYER_ALUCARD) ||
+            (g_DemoMode != Demo_None)) {
+            if (!self->params) {
+                DestroyEntity(self);
+                break;
+            }
+        }
+        if (self->params) {
+            ptr = SetCutsceneScript(D_80181B65);
+        } else {
+            ptr = SetCutsceneScript(D_801816C8);
+        }
+        if (ptr) {
+            self->flags |= FLAG_HAS_PRIMS | FLAG_UNK_2000;
             g_CutsceneFlags = 0;
-            D_801C2580 = 0;
-            D_801C24C8 = 0;
+            D_801A3F88 = 0;
+            skip_cutscene = 0;
             D_8003C704 = 1;
             self->primIndex = g_Dialogue.primIndex[2];
-            self->step += 1;
-            if (g_DemoMode != Demo_None) {
-                D_801C24C8 = 1;
-                D_801C2580 = 1;
-                g_api.FreePrimitives(self->primIndex);
-                self->flags &= ~FLAG_HAS_PRIMS;
-            }
+            self->step++;
+        }
+        if ((g_CastleFlags[SUCC_CS_DONE] != 0) ||
+            (g_PlayableCharacter != PLAYER_ALUCARD) ||
+            (g_DemoMode != Demo_None)) {
+            skip_cutscene = 1;
         }
         break;
     case 1:
         // this is a huge While-loop! Don't miss it!
         while (1) {
-            if ((g_Dialogue.nextCharTimer != 0) && (D_801C24C8 == 0)) {
+            if ((g_Dialogue.nextCharTimer != 0) && !skip_cutscene) {
                 g_Dialogue.nextCharTimer--;
                 return;
             }
             nextChar = *g_Dialogue.nextCharDialogue++;
             switch (nextChar) {
-            case 0:
+            case CSOP_END_CUTSCENE:
                 self->step = 7;
                 return;
-            case 1:
-                if (D_801C24C8 != 0) {
+            case CSOP_LINE_BREAK:
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharX = g_Dialogue.nextLineX;
@@ -172,17 +191,17 @@ void EntityDraculaCutscene(Entity* self) {
                 self->step_s = 0;
                 self->step++;
                 return;
-            case 2:
+            case CSOP_SET_SPEED:
                 g_Dialogue.unk17 = *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 3:
+            case CSOP_SET_WAIT:
                 g_Dialogue.nextCharTimer = *g_Dialogue.nextCharDialogue++;
-                if (D_801C24C8 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 return;
-            case 4:
-                if (D_801C24C8 != 0) {
+            case CSOP_SCRIPT_UNKNOWN_4:
+                if (skip_cutscene) {
                     continue;
                 }
                 prim = g_Dialogue.prim[0];
@@ -191,8 +210,8 @@ void EntityDraculaCutscene(Entity* self) {
                     prim = prim->next;
                 }
                 return;
-            case 5:
-                if (D_801C24C8 != 0) {
+            case CSOP_SET_PORTRAIT:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 2;
                     continue;
                 }
@@ -200,9 +219,9 @@ void EntityDraculaCutscene(Entity* self) {
                 i = *g_Dialogue.nextCharDialogue++;
                 nextChar2 = *g_Dialogue.nextCharDialogue++;
                 prim = g_Dialogue.prim[5];
-                uCoord = D_80180784[nextChar2 & 1]; // a1
-                vCoord = D_80180788[nextChar2 & 1]; // a2
-                prim->clut = D_8018078C[i];
+                uCoord = D_8018088C[nextChar2 & 1];
+                vCoord = D_80180890[nextChar2 & 1];
+                prim->clut = D_80180894[i];
                 prim->tpage = 0x90;
                 if (nextChar2 & 0x80) {
                     prim->u0 = prim->u2 = uCoord + 0x2F;
@@ -217,7 +236,7 @@ void EntityDraculaCutscene(Entity* self) {
                     g_Dialogue.startX - 0x1E;
                 prim->y0 = prim->y1 = prim->y2 = prim->y3 =
                     g_Dialogue.startY + 0x24;
-                g_Dialogue.clutIndex = D_80180794[i];
+                g_Dialogue.clutIndex = D_801808A0[i];
                 CutsceneUnk1();
                 CutsceneUnk4();
                 prim->priority = 0x1FE;
@@ -226,8 +245,8 @@ void EntityDraculaCutscene(Entity* self) {
                 g_Dialogue.portraitAnimTimer = 6;
                 self->step = 3;
                 return;
-            case 6:
-                if (D_801C24C8 != 0) {
+            case CSOP_NEXT_DIALOG:
+                if (skip_cutscene) {
                     continue;
                 }
                 for (prim = g_Dialogue.prim[0], i = 0; i < 5; i++) {
@@ -239,8 +258,8 @@ void EntityDraculaCutscene(Entity* self) {
                 g_Dialogue.portraitAnimTimer = 6;
                 self->step = 4;
                 return;
-            case 7:
-                if (D_801C24C8 != 0) {
+            case CSOP_SET_POS:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -258,16 +277,15 @@ void EntityDraculaCutscene(Entity* self) {
                 self->step = 5;
                 self->step_s = 0;
                 return;
-
-            case 8:
-                if (D_801C24C8 != 0) {
+            case CSOP_CLOSE_DIALOG:
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.portraitAnimTimer = 0x18;
                 self->step = 6;
                 return;
-            case 9:
-                if (D_801C24C8 != 0) {
+            case CSOP_PLAY_SOUND:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -277,8 +295,8 @@ void EntityDraculaCutscene(Entity* self) {
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            case 10:
-                if (D_801C24C8 != 0) {
+            case CSOP_WAIT_FOR_SOUND:
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != false) {
@@ -286,8 +304,8 @@ void EntityDraculaCutscene(Entity* self) {
                 }
                 *g_Dialogue.nextCharDialogue--;
                 return;
-            case 11:
-                if (D_801C24C8 != 0) {
+            case CSOP_UNK_11:
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != true) {
@@ -295,52 +313,49 @@ void EntityDraculaCutscene(Entity* self) {
                 }
                 *g_Dialogue.nextCharDialogue--;
                 return;
-            case 12:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                CutsceneUnk6((u8*)bit_shifty);
+            case CSOP_SCRIPT_UNKNOWN_12:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                CutsceneUnk6((u8*)ptr);
                 continue;
-            case 13:
+            case CSOP_SCRIPT_UNKNOWN_13:
                 continue;
-            case 14:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                // This needs help. Casting the const to short is good.
-                bit_shifty += (u16)0x100000;
-                g_Dialogue.nextCharDialogue += *(u16*)bit_shifty << 2;
+            case CSOP_SCRIPT_UNKNOWN_14:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr += 0x100000;
+                g_Dialogue.nextCharDialogue += *(u8*)ptr << 2;
 
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
-
-            case 15: // c64
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+            case CSOP_SCRIPT_UNKNOWN_15:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
-
-            case 16:
+            case CSOP_WAIT_FOR_FLAG:
                 if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
                     g_Dialogue.nextCharDialogue--;
                     return;
@@ -348,46 +363,60 @@ void EntityDraculaCutscene(Entity* self) {
                 g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue);
                 *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 17:
+            case CSOP_SET_FLAG:
                 g_CutsceneFlags |= 1 << *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 18:
+            case CSOP_SCRIPT_UNKNOWN_18:
                 g_Dialogue.unk3C = 0;
                 continue;
-            case 19:
-                if (D_801C24C8 != 0) {
+            case CSOP_LOAD_PORTRAIT:
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 5;
                 } else {
-                    bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;  // 8e4
-                    bit_shifty <<= 4;                                  // 918
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++; // 91c
-                    bit_shifty <<= 4;                                  // 920
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++; // 924
-                    bit_shifty <<= 4;                                  // 928
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++; // 92c
-                    bit_shifty += 0x100000;                            //
+                    ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr += 0x100000;
                     nextChar2 = g_Dialogue.nextCharDialogue++[0];
-                    LoadTPage((u32*)bit_shifty, 1, 0, D_80180790[nextChar2],
-                              0x100, 0x30, 0x48);
+                    LoadTPage((u32*)ptr, 1, 0, D_8018089C[nextChar2], 0x100,
+                              0x30, 0x48);
                 }
                 continue;
-            case 20:
+            case CSOP_SCRIPT_UNKNOWN_20:
                 nextChar = *g_Dialogue.nextCharDialogue++;
                 nextChar <<= 4;
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            // Note: Other cutscenes have Case 21-24 here, this one is missing
-            // it.
+            case CSOP_SCRIPT_UNKNOWN_21:
+                g_CutsceneFlags = 0;
+                skip_cutscene = 0;
+                D_801A3F88 = 0;
+                continue;
+            case CSOP_SCRIPT_UNKNOWN_22:
+                g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue++);
+                continue;
+            case CSOP_SCRIPT_UNKNOWN_23:
+                return;
+            case CSOP_SCRIPT_UNKNOWN_24:
+                if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
+                    *g_Dialogue.nextCharDialogue--;
+                    return;
+                }
+                *g_Dialogue.nextCharDialogue++;
+                continue;
             default:
-                if (D_801C24C8 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharTimer = g_Dialogue.unk17;
             }
             break; // This breaks the big while-loop!
         }
-        // End of while loop, comes at a58
         if (nextChar == 0x20) {
             g_Dialogue.nextCharX += 2;
             return;
@@ -413,10 +442,6 @@ void EntityDraculaCutscene(Entity* self) {
         prim->x1 = prim->x3 += 4;
         prim->y0 = prim->y1 -= 6;
         prim->y2 = prim->y3 += 6;
-        // This block is in case 4 for other cutscenes, case 3 for this one
-        if (prim->x1 >= (g_Dialogue.startX - 2)) {
-            prim->x1 = prim->x3 = g_Dialogue.startX - 3;
-        }
         g_Dialogue.portraitAnimTimer--;
         if (!g_Dialogue.portraitAnimTimer) {
             self->step = 1;
@@ -432,6 +457,9 @@ void EntityDraculaCutscene(Entity* self) {
         prim->x1 = prim->x3 -= 4;
         prim->y0 = prim->y1 += 6;
         prim->y2 = prim->y3 -= 6;
+        if (prim->x1 >= (g_Dialogue.startX - 2)) {
+            prim->x1 = prim->x3 = g_Dialogue.startX - 3;
+        }
         g_Dialogue.portraitAnimTimer--;
         if (!g_Dialogue.portraitAnimTimer) {
             self->step = 1;
@@ -441,7 +469,7 @@ void EntityDraculaCutscene(Entity* self) {
     case 5:
         switch (self->step_s) {
         case 0:
-            D_801C2580 = 1;
+            D_801A3F88 = 1;
             primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 0x48);
             if (primIndex == -1) {
                 DestroyEntity(self);
@@ -456,14 +484,14 @@ void EntityDraculaCutscene(Entity* self) {
                 prim->y0 = prim->y1 = g_Dialogue.startY + j;
                 prim->priority = 0x1FE;
                 prim->drawMode = DRAW_DEFAULT;
-                prim->x2 = D_80180798[j];
+                prim->x2 = D_801808A8[j];
                 prim->x3 = 0xF70;
 
                 j++;
             }
             self->step_s++;
             return;
-        case 1: /* switch 3 */
+        case 1:
             for (j = 0, prim = &g_PrimBuf[g_Dialogue.primIndex[0]];
                  prim != NULL; prim = prim->next) {
                 prim->x3 -= prim->x2;
@@ -541,8 +569,13 @@ void EntityDraculaCutscene(Entity* self) {
         break;
 
     case 7:
+        if (self->params) {
+            g_CastleFlags[SUCC_CS_DONE] = 1;
+            g_Settings.D_8003CB04 |= 0x2000;
+        } else {
+            g_Settings.D_8003CB04 |= 0x1000;
+        }
         D_8003C704 = 0;
-        g_Settings.D_8003CB04 |= 1;
         DestroyEntity(self);
         break;
     }
