+++ src/st/dre/dre.h
@@ -131,7 +131,7 @@ extern u8 g_ESoulStealOrbAnim[];
 
 // *** EntitySoulStealOrb properties END ***
 
-extern s32 D_801A3ED4;
+extern s32 skip_cutscene;
 extern Primitive* D_801A3EF0[];
 extern s32 D_801A3F08;
 extern s32 D_801A3F0C;
@@ -147,7 +147,7 @@ extern u16 D_801808A0[];
 extern s16 D_801808A8[];
 extern const char D_80181B65[];
 extern const char D_801816C8[];
-extern s32 D_801A3ED4;
+extern s32 skip_cutscene;
 extern u32 g_CutsceneFlags;
 extern s32 D_801A3F88;
