+++ src/st/no3/cutscene.c
@@ -1,120 +1,74 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-#include "cen.h"
+#include "no3.h"
 #include <cutscene.h>
 
-static u8 __unused[0xC00];
-static s32 D_8019D374;
-static Dialogue g_Dialogue;
-static u32 __unused0[26];
-u32 g_CutsceneFlags;
-u32 D_8019D428;
-
-// Bizarre variable - u8 here, but u16 in EntityHeartDrop
-// Possible that since EntityHeartDrop is not used in CEN, that data
-// is reused here for the holy glasses cutscene? Very unusual.
-u8 g_HeartDropArray[] = {0, 64, 0, 0};
-static u8 D_801805DC[] = {0, 0, 0, 0};
-
-// clut
-static u16 D_801805E0[] = {
-    LAYER_CLUT_ALT | 0,
-    LAYER_CLUT_ALT | 8,
-    LAYER_CLUT_ALT | 16,
-    LAYER_CLUT_ALT | 24,
-};
-static u16 D_801805E8[] = {0, 32};
-
-static u16 D_801805EC[] = {
-    0x01A1,
-    0x01A1,
-    0x01A1,
-    0x01A1,
-};
-
-// x-coords
-static s16 D_801805F4[] = {
-    0x08, 0x13, 0x11, 0x31, 0x4F, 0x26, 0x36, 0x1D, 0x1B, 0x33, 0x2C, 0x21,
-    0x19, 0x0A, 0x33, 0x1F, 0x48, 0x2F, 0x13, 0x19, 0x4D, 0x4B, 0x17, 0x1D,
-    0x12, 0x02, 0x1B, 0x2A, 0x50, 0x45, 0x32, 0x0D, 0x2A, 0x4D, 0x06, 0x27,
-    0x07, 0x48, 0x2F, 0x1B, 0x36, 0x22, 0x39, 0x14, 0x39, 0x1D, 0x0A, 0x35,
-    0x10, 0x1B, 0x3D, 0x17, 0x2E, 0x0B, 0x49, 0x42, 0x3D, 0x2A, 0x01, 0x0C,
-    0x1B, 0x34, 0x41, 0x35, 0x08, 0x0E, 0x4D, 0x11, 0x34, 0x41, 0x29, 0x48,
-};
-
-static const char* g_ActorNames[] = {
-    _S("Alucard"),
-    _S("Maria"),
-};
-
-static const char _pad[4] = "";
-
-#include "../cutscene_unk1.h"
+// similar but not an exact duplicate
+static void CutsceneUnk1(void) {
+    g_Dialogue.nextLineX = 0x182; // Note that these two lines are "= 2"
+    g_Dialogue.nextCharX = 0x182; // for all other cutscenes (so far)!
+    g_Dialogue.nextCharY = 0;
+    g_Dialogue.unk12 = 0;
+    g_Dialogue.nextCharTimer = 0;
+    g_Dialogue.unk17 = 8;
+    g_Dialogue.nextLineY = g_Dialogue.startY + 0x14;
+}
 
 #include "../set_cutscene_script.h"
 
-#include "../cutscene_unk3.h"
+static void CutsceneUnk3(s16 yOffset) {
+    RECT rect;
 
-#include "../cutscene_unk4.h"
+    rect.x = 384;
+    rect.y = (yOffset * 12) + 256;
+    rect.w = 64;
+    rect.h = 12;
+    ClearImage(&rect, 0, 0, 0);
+}
+
+static void CutsceneUnk4(void) {
+    Primitive* prim;
+
+    CutsceneUnk3(g_Dialogue.nextCharY);
+    prim = g_Dialogue.prim[g_Dialogue.nextCharY];
+    prim->tpage = 0x16;
+    prim->clut = g_Dialogue.clutIndex;
+    prim->y0 = g_Dialogue.nextLineY;
+    prim->u0 = 0;
+    prim->x0 = g_Dialogue.startX;
+    prim->x0 = prim->x0 + 4;
+    prim->v0 = g_Dialogue.nextCharY * 0xC;
+    prim->u1 = 0xC0;
+    prim->v1 = 0xC;
+    prim->priority = 0x1FF;
+    prim->drawMode = DRAW_DEFAULT;
+}
 
 #include "../cutscene_avatar.h"
 
 #include "../cutscene_unk6.h"
 
-void CutsceneRun(void) {
-    Entity* entity;
-    u16 startTimer;
+#include "../cutscene_run.h"
 
-    g_Dialogue.timer++;
-    // protect from overflows
-    if (g_Dialogue.timer >= 0xFFFF) {
-        g_Dialogue.unk3C = 0;
-        return;
-    }
-    while (true) {
-        // Start the dialogue script only if the start timer has passed
-        startTimer = *g_Dialogue.unk40++ << 8;
-        startTimer |= *g_Dialogue.unk40++;
-        if (g_Dialogue.timer < startTimer) {
-            // Re-evaluate the condition at the next frame
-            g_Dialogue.unk40 -= 2;
-            return;
+static void func_801BECCC(Entity* self) {
+    if (g_pads[0].tapped == PAD_START) {
+        skip_cutscene = 1;
+        g_api.FreePrimitives(self->primIndex);
+        self->flags ^= FLAG_HAS_PRIMS;
+        if (g_Dialogue.primIndex[1] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[1]);
         }
-        switch (*g_Dialogue.unk40++) {
-        case 0:
-            entity =
-                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
-            DestroyEntity(entity);
-            entity->entityId = *g_Dialogue.unk40++;
-            entity->pfnUpdate = PfnEntityUpdates[entity->entityId - 1];
-            entity->posX.i.hi = *g_Dialogue.unk40++ * 0x10;
-            entity->posX.i.hi |= *g_Dialogue.unk40++;
-            entity->posY.i.hi = *g_Dialogue.unk40++ * 0x10;
-            entity->posY.i.hi |= *g_Dialogue.unk40++;
-            entity->posX.i.hi -= g_Tilemap.scrollX.i.hi;
-            entity->posY.i.hi -= g_Tilemap.scrollY.i.hi;
-            break;
-        case 1:
-            entity =
-                &g_Entities[*g_Dialogue.unk40++ & 0xFF] + STAGE_ENTITY_START;
-            DestroyEntity(entity);
-            break;
-        case 2:
-            if (!((g_CutsceneFlags >> *g_Dialogue.unk40) & 1)) {
-                g_Dialogue.unk40--;
-                return;
-            }
-            g_CutsceneFlags &= ~(1 << *g_Dialogue.unk40++);
-            break;
-        case 3:
-            g_CutsceneFlags |= 1 << *g_Dialogue.unk40++;
-            break;
+        if (g_Dialogue.primIndex[0] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[0]);
         }
+        g_api.PlaySfx(SET_STOP_MUSIC);
+        self->step = 1;
+        self->step_s = 0;
     }
 }
 
 #include "../cutscene_scale_avatar.h"
 
-void EntityHolyGlassesCutscene(Entity* self) {
+void NO3_CutsceneExec(Entity* self) {
     RECT rect;
     Primitive* prim;
     s32 primIndex;
@@ -124,42 +78,33 @@ void EntityHolyGlassesCutscene(Entity* self) {
     s16 vCoord;
     u16 nextChar;
     s32 nextChar2;
-    u32 bit_shifty;
+    u_long ptr;
 
-    if (self->step != 0) {
-        if ((D_8019D428 != 0) && (D_8019D374 == 0) &&
-            ((g_Settings.D_8003CB04 & 0x800) ||
-             (g_IsTimeAttackUnlocked != 0)) &&
-            (g_pads[0].tapped & PAD_START)) {
-            D_8019D374 = 1;
-            g_api.FreePrimitives(self->primIndex);
-            self->flags ^= FLAG_HAS_PRIMS;
-            if (g_Dialogue.primIndex[1] != -1) {
-                g_api.FreePrimitives(g_Dialogue.primIndex[1]);
-            }
-            if (g_Dialogue.primIndex[0] != -1) {
-                g_api.FreePrimitives(g_Dialogue.primIndex[0]);
-            }
-            g_api.PlaySfx(SET_STOP_MUSIC);
-            self->step = 1;
-            self->step_s = 0;
+    if (self->step) {
+        if ((D_801D7DD4 != 0) && !skip_cutscene &&
+            ((g_IsTimeAttackUnlocked) || (g_Settings.D_8003CB04 & 4))) {
+            func_801BECCC(self);
         }
-        if ((self->step) && (g_Dialogue.unk3C != 0)) {
+        if ((self->step != 0) && (D_801D7D60 != 0)) {
             CutsceneRun();
         }
     }
     switch (self->step) {
     case 0:
-        if (g_CastleFlags[HG_CS_DONE] != 0) {
+        CreateEntityFromCurrentEntity(E_BG_LIGHTNING, &g_Entities[192]);
+        if ((g_CastleFlags[CASTLE_FLAG_53]) ||
+            (g_PlayableCharacter != PLAYER_ALUCARD)) {
             DestroyEntity(self);
+            g_Entities[192].params = 0;
             return;
         }
-        if (SetCutsceneScript(g_CutsceneScript)) {
+        g_Entities[192].params = 0x100;
+        if (SetCutsceneScript(D_80184CE0)) {
             self->flags |= FLAG_HAS_PRIMS | FLAG_UNK_2000;
-            D_8003C704 = 1;
             g_CutsceneFlags = 0;
-            D_8019D428 = 0;
-            D_8019D374 = 0;
+            D_801D7DD4 = 0;
+            skip_cutscene = 0;
+            D_8003C704 = 1;
             self->primIndex = g_Dialogue.primIndex[2];
             self->step++;
         }
@@ -167,7 +112,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
     case 1:
         // this is a huge While-loop! Don't miss it!
         while (1) {
-            if ((g_Dialogue.nextCharTimer != 0) && (D_8019D374 == 0)) {
+            if ((g_Dialogue.nextCharTimer != 0) && !skip_cutscene) {
                 g_Dialogue.nextCharTimer--;
                 return;
             }
@@ -177,7 +122,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 self->step = 7;
                 return;
             case CSOP_LINE_BREAK:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharX = g_Dialogue.nextLineX;
@@ -208,12 +153,12 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 continue;
             case CSOP_SET_WAIT:
                 g_Dialogue.nextCharTimer = *g_Dialogue.nextCharDialogue++;
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 return;
-            case 4:
-                if (D_8019D374 != 0) {
+            case CSOP_SCRIPT_UNKNOWN_4:
+                if (skip_cutscene) {
                     continue;
                 }
                 prim = g_Dialogue.prim[0];
@@ -223,7 +168,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 }
                 return;
             case CSOP_SET_PORTRAIT:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 2;
                     continue;
                 }
@@ -231,10 +176,10 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 i = *g_Dialogue.nextCharDialogue++;
                 nextChar2 = *g_Dialogue.nextCharDialogue++;
                 prim = g_Dialogue.prim[5];
-                uCoord = g_HeartDropArray[nextChar2 & 1];
-                vCoord = D_801805DC[nextChar2 & 1];
-                prim->clut = D_801805E0[i];
-                prim->tpage = 0x90;
+                uCoord = D_80181A28[nextChar2 & 1];
+                vCoord = D_80181A2C[nextChar2 & 1];
+                prim->clut = D_80181A30[i];
+                prim->tpage = 0x94; // Weird, this is 0x90 in other cutscenes
                 if (nextChar2 & 0x80) {
                     prim->u0 = prim->u2 = uCoord + 0x2F;
                     prim->u1 = prim->u3 = uCoord;
@@ -248,7 +193,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                     g_Dialogue.startX - 0x1E;
                 prim->y0 = prim->y1 = prim->y2 = prim->y3 =
                     g_Dialogue.startY + 0x24;
-                g_Dialogue.clutIndex = D_801805EC[i];
+                g_Dialogue.clutIndex = D_80181A38[i];
                 CutsceneUnk1();
                 CutsceneUnk4();
                 prim->priority = 0x1FE;
@@ -258,7 +203,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 self->step = 3;
                 return;
             case CSOP_NEXT_DIALOG:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 for (prim = g_Dialogue.prim[0], i = 0; i < 5; i++) {
@@ -271,7 +216,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 self->step = 4;
                 return;
             case CSOP_SET_POS:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -290,14 +235,14 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 self->step_s = 0;
                 return;
             case CSOP_CLOSE_DIALOG:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.portraitAnimTimer = 0x18;
                 self->step = 6;
                 return;
             case CSOP_PLAY_SOUND:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -308,7 +253,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 g_api.PlaySfx(nextChar);
                 continue;
             case CSOP_WAIT_FOR_SOUND:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != false) {
@@ -317,7 +262,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 *g_Dialogue.nextCharDialogue--;
                 return;
             case CSOP_UNK_11:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != true) {
@@ -325,47 +270,47 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 }
                 *g_Dialogue.nextCharDialogue--;
                 return;
-            case 12:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                CutsceneUnk6((u8*)bit_shifty);
+            case CSOP_SCRIPT_UNKNOWN_12:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                CutsceneUnk6((u8*)ptr);
                 continue;
-            case 13:
+            case CSOP_SCRIPT_UNKNOWN_13:
                 continue;
-            case 14:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty += (u16)0x100000;
-                g_Dialogue.nextCharDialogue += *(u16*)bit_shifty << 2;
+            case CSOP_SCRIPT_UNKNOWN_14:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr += (u16)0x100000;
+                g_Dialogue.nextCharDialogue += *(u16*)ptr << 2;
 
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
-            case 15:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+            case CSOP_SCRIPT_UNKNOWN_15:
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
             case CSOP_WAIT_FOR_FLAG:
                 if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
@@ -375,27 +320,27 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue);
                 *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 17:
+            case CSOP_SET_FLAG:
                 g_CutsceneFlags |= 1 << *g_Dialogue.nextCharDialogue++;
                 continue;
-            case 18:
+            case CSOP_SCRIPT_UNKNOWN_18:
                 g_Dialogue.unk3C = 0;
                 continue;
             case CSOP_LOAD_PORTRAIT:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 5;
                 } else {
-                    bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty += 0x100000;
+                    ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr <<= 4;
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                    ptr += 0x100000;
                     nextChar2 = g_Dialogue.nextCharDialogue++[0];
-                    LoadTPage((u_long*)bit_shifty, 1, 0, D_801805E8[nextChar2],
-                              0x100, 48, 72);
+                    LoadTPage((u32*)ptr, 1, 0, D_80181A34[nextChar2], 0x100,
+                              0x30, 0x48);
                 }
                 continue;
             case CSOP_SCRIPT_UNKNOWN_20:
@@ -404,17 +349,17 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            case 21:
+            case CSOP_SCRIPT_UNKNOWN_21:
                 g_CutsceneFlags = 0;
-                D_8019D374 = 0;
-                D_8019D428 = 0;
+                skip_cutscene = 0;
+                D_801D7DD4 = 0;
                 continue;
-            case 22:
+            case CSOP_SCRIPT_UNKNOWN_22:
                 g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue++);
                 continue;
-            case 23:
+            case CSOP_SCRIPT_UNKNOWN_23:
                 return;
-            case 24:
+            case CSOP_SCRIPT_UNKNOWN_24:
                 if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
                     *g_Dialogue.nextCharDialogue--;
                     return;
@@ -422,7 +367,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 *g_Dialogue.nextCharDialogue++;
                 continue;
             default:
-                if (D_8019D374 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharTimer = g_Dialogue.unk17;
@@ -437,7 +382,8 @@ void EntityHolyGlassesCutscene(Entity* self) {
         rect.y = ((u32)(nextChar & 0xF0) >> 1) + 0xF0;
         rect.w = 2;
         rect.h = 8;
-        vCoord = (g_Dialogue.nextCharY * 0xC) + 0x180;
+        // Other cutscenes have + 0x180 here
+        vCoord = (g_Dialogue.nextCharY * 0xC) + 0x100;
         MoveImage(&rect, g_Dialogue.nextCharX, vCoord);
         g_Dialogue.nextCharX += 2;
         break;
@@ -481,7 +427,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
     case 5:
         switch (self->step_s) {
         case 0:
-            D_8019D428 = 1;
+            D_801D7DD4 = 1;
             primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 0x48);
             if (primIndex == -1) {
                 DestroyEntity(self);
@@ -496,7 +442,7 @@ void EntityHolyGlassesCutscene(Entity* self) {
                 prim->y0 = prim->y1 = g_Dialogue.startY + j;
                 prim->priority = 0x1FE;
                 prim->drawMode = DRAW_DEFAULT;
-                prim->x2 = D_801805F4[j];
+                prim->x2 = D_80181A3C[j];
                 prim->x3 = 0xF70;
 
                 j++;
@@ -581,12 +527,12 @@ void EntityHolyGlassesCutscene(Entity* self) {
         break;
 
     case 7:
-        DestroyEntity(self);
-        g_CastleFlags[HG_CS_DONE] = 1;
+        g_CastleFlags[CASTLE_FLAG_53] = 1;
         g_api.TimeAttackController(
-            TIMEATTACK_EVENT_GET_HOLYGLASSES, TIMEATTACK_SET_RECORD);
+            TIMEATTACK_EVENT_MEET_DEATH, TIMEATTACK_SET_RECORD);
         D_8003C704 = 0;
-        g_Settings.D_8003CB04 |= 0x800;
+        g_Settings.D_8003CB04 |= 4;
+        DestroyEntity(self);
         break;
     }
 }
