+++ src/st/nz0/bss.c
@@ -2,7 +2,7 @@
 #include <common.h>
 
 STATIC_PAD_BSS(3072);
-u32 D_801CB684;
+u32 skip_cutscene;
 u32 g_Dialogue;
 u16 D_801CB68C;
 u16 D_801CB68E;
