+++ src/st/sel/bss.c
@@ -53,7 +53,7 @@ u32 D_801BC340;
 u32 g_StreamDiskIsReady;
 u32 D_801BC348;
 u32 D_801BC34C;
-u32 D_801BC350;
+u32 skip_cutscene;
 u32 g_Dialogue;
 u16 D_801BC358;
 u16 D_801BC35A;
