+++ src/st/st0/bss.c
@@ -8,7 +8,7 @@ u16 D_801C1BC8;
 u16 D_801C1BCA;
 u16 D_801C1BCC[1022];
 u32 D_801C23C8[64];
-u32 D_801C24C8;
+u32 skip_cutscene;
 u32 g_Dialogue;
 u16 D_801C24D0;
 u16 D_801C24D2;
