+++ src/st/st0/cutscene.c
@@ -1,28 +1,84 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-#include "nz0.h"
-
-// This file consists of EntityMariaCutscene, and all the functions it calls.
-// No function in NZ0 calls anything in there, nor does anything in here call
-// any NZ0 functions, besides the ones in this file.
+#include "st0.h"
+#include <cutscene.h>
 
 #include "../cutscene_unk1.h"
 
-#include "../set_cutscene_script.h"
+static u8 SetCutsceneScript(const char* textDialogue) {
+    Primitive* prim;
+    s16 firstPrimIndex;
+
+    firstPrimIndex = g_api.AllocPrimitives(PRIM_SPRT, 8);
+    g_Dialogue.primIndex[2] = firstPrimIndex;
+    if (firstPrimIndex == -1) {
+        g_Dialogue.primIndex[2] = 0;
+        return 0;
+    }
+    g_Dialogue.nextCharDialogue = textDialogue;
+    g_Dialogue.unk3C = 0;
+    g_Dialogue.primIndex[1] = -1;
+    g_Dialogue.primIndex[0] = -1;
+    CutsceneUnk1();
+
+    if (prim && prim) { // !FAKE
+    }
+
+    prim = g_Dialogue.prim[0] = &g_PrimBuf[g_Dialogue.primIndex[2]];
+
+    prim->drawMode = DRAW_HIDE;
+    prim = g_Dialogue.prim[1] = prim->next;
+
+    prim->drawMode = DRAW_HIDE;
+    prim = g_Dialogue.prim[2] = prim->next;
+
+    prim->drawMode = DRAW_HIDE;
+    prim = g_Dialogue.prim[3] = prim->next;
+
+    prim->drawMode = DRAW_HIDE;
+    prim = g_Dialogue.prim[4] = prim->next;
+
+    prim->drawMode = DRAW_HIDE;
+    prim = g_Dialogue.prim[5] = prim->next;
+
+    prim->type = 4;
+    prim->drawMode = DRAW_HIDE;
+
+    prim = prim->next;
+    prim->type = 3;
+    prim->r0 = prim->r1 = prim->r2 = prim->r3 = 0xFF;
+    prim->g0 = prim->g1 = prim->g2 = prim->g3 = 0;
+    prim->b0 = prim->b1 = prim->b2 = prim->b3 = 0;
+    prim->x0 = prim->x2 = 4;
+    prim->x1 = prim->x3 = 0xF8;
+    prim->priority = 0x1FD;
+    prim->drawMode = DRAW_HIDE;
+
+    prim = prim->next;
+    prim->type = 1;
+    prim->x0 = 3;
+    prim->y0 = 0x2F;
+    prim->v0 = 0x4A;
+    prim->r0 = prim->g0 = prim->b0 = 0xFF;
+    prim->priority = 0x1FC;
+    prim->drawMode = DRAW_HIDE;
+    return 1;
+}
 
 #include "../cutscene_unk3.h"
 
 #include "../cutscene_unk4.h"
 
+const char* g_ActorNames[] = {_S("Richter"), _S("Dracula")};
+
 #include "../cutscene_avatar.h"
 
 #include "../cutscene_unk6.h"
 
-#include "../cutscene.h"
+#include "../cutscene_run.h"
 
 #include "../cutscene_scale_avatar.h"
 
-// cutscene where alucard and maria discuss castle changing
-void EntityMariaCutscene(Entity* self) {
+void ST0_CutsceneExec(Entity* self) {
     RECT rect;
     Primitive* prim;
     s32 primIndex;
@@ -32,51 +88,50 @@ void EntityMariaCutscene(Entity* self) {
     s16 vCoord;
     u16 nextChar;
     s32 nextChar2;
-    s32 bit_shifty;
+    u_long ptr;
 
-    if (self->step != 0) {
-        if ((D_801CB73C != 0) && (D_801CB684 == 0) &&
-            ((g_Settings.D_8003CB04 & 0x100) ||
-             (g_IsTimeAttackUnlocked != 0)) &&
-            (g_pads[0].tapped == PAD_START)) {
-            D_801CB684 = 1;
-            g_api.FreePrimitives(self->primIndex);
-            self->flags ^= FLAG_HAS_PRIMS;
-            if (g_Dialogue.primIndex[1] != -1) {
-                g_api.FreePrimitives(g_Dialogue.primIndex[1]);
-            }
-            if (g_Dialogue.primIndex[0] != -1) {
-                g_api.FreePrimitives(g_Dialogue.primIndex[0]);
-            }
-            g_api.PlaySfx(SET_STOP_MUSIC);
-            self->step = 1;
-            self->step_s = 0;
+    if ((D_801C2580 != 0) && !skip_cutscene &&
+        ((g_Settings.D_8003CB04 & 1) || (g_IsTimeAttackUnlocked != 0)) &&
+        (g_pads[0].tapped == PAD_START)) {
+        skip_cutscene = 1;
+        g_api.FreePrimitives(self->primIndex);
+        self->flags &= ~FLAG_HAS_PRIMS;
+        if (g_Dialogue.primIndex[1] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[1]);
         }
-        if ((self->step) && (g_Dialogue.unk3C != 0)) {
-            CutsceneRun();
+        if (g_Dialogue.primIndex[0] != -1) {
+            g_api.FreePrimitives(g_Dialogue.primIndex[0]);
         }
+        g_api.PlaySfx(0x80);
+        self->step = 1;
+        self->step_s = 0;
     }
+    if (self->step && (g_Dialogue.unk3C != 0)) {
+        CutsceneRun();
+    }
+
     switch (self->step) {
     case 0:
-        if ((g_CastleFlags[CASTLE_FLAG_133] != 0) ||
-            (g_PlayableCharacter != PLAYER_ALUCARD)) {
-            DestroyEntity(self);
-            return;
-        }
-        if (SetCutsceneScript(D_80183B0C)) {
-            self->flags |= FLAG_HAS_PRIMS | FLAG_UNK_2000;
+        if (SetCutsceneScript(D_801829D8)) {
+            self->flags |= FLAG_HAS_PRIMS;
             g_CutsceneFlags = 0;
-            D_801CB73C = 0;
-            D_801CB684 = 0;
+            D_801C2580 = 0;
+            skip_cutscene = 0;
             D_8003C704 = 1;
             self->primIndex = g_Dialogue.primIndex[2];
-            self->step++;
+            self->step += 1;
+            if (g_DemoMode != Demo_None) {
+                skip_cutscene = 1;
+                D_801C2580 = 1;
+                g_api.FreePrimitives(self->primIndex);
+                self->flags &= ~FLAG_HAS_PRIMS;
+            }
         }
         break;
     case 1:
         // this is a huge While-loop! Don't miss it!
         while (1) {
-            if ((g_Dialogue.nextCharTimer != 0) && (D_801CB684 == 0)) {
+            if ((g_Dialogue.nextCharTimer != 0) && !skip_cutscene) {
                 g_Dialogue.nextCharTimer--;
                 return;
             }
@@ -86,7 +141,7 @@ void EntityMariaCutscene(Entity* self) {
                 self->step = 7;
                 return;
             case 1:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharX = g_Dialogue.nextLineX;
@@ -117,12 +172,12 @@ void EntityMariaCutscene(Entity* self) {
                 continue;
             case 3:
                 g_Dialogue.nextCharTimer = *g_Dialogue.nextCharDialogue++;
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 return;
             case 4:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 prim = g_Dialogue.prim[0];
@@ -132,7 +187,7 @@ void EntityMariaCutscene(Entity* self) {
                 }
                 return;
             case 5:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 2;
                     continue;
                 }
@@ -140,9 +195,9 @@ void EntityMariaCutscene(Entity* self) {
                 i = *g_Dialogue.nextCharDialogue++;
                 nextChar2 = *g_Dialogue.nextCharDialogue++;
                 prim = g_Dialogue.prim[5];
-                uCoord = D_801813C8[nextChar2 & 1];
-                vCoord = D_801813CC[nextChar2 & 1];
-                prim->clut = D_801813D0[i];
+                uCoord = D_80180784[nextChar2 & 1]; // a1
+                vCoord = D_80180788[nextChar2 & 1]; // a2
+                prim->clut = D_8018078C[i];
                 prim->tpage = 0x90;
                 if (nextChar2 & 0x80) {
                     prim->u0 = prim->u2 = uCoord + 0x2F;
@@ -157,7 +212,7 @@ void EntityMariaCutscene(Entity* self) {
                     g_Dialogue.startX - 0x1E;
                 prim->y0 = prim->y1 = prim->y2 = prim->y3 =
                     g_Dialogue.startY + 0x24;
-                g_Dialogue.clutIndex = D_801813D8[i];
+                g_Dialogue.clutIndex = D_80180794[i];
                 CutsceneUnk1();
                 CutsceneUnk4();
                 prim->priority = 0x1FE;
@@ -167,7 +222,7 @@ void EntityMariaCutscene(Entity* self) {
                 self->step = 3;
                 return;
             case 6:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 for (prim = g_Dialogue.prim[0], i = 0; i < 5; i++) {
@@ -180,7 +235,7 @@ void EntityMariaCutscene(Entity* self) {
                 self->step = 4;
                 return;
             case 7:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -200,14 +255,14 @@ void EntityMariaCutscene(Entity* self) {
                 return;
 
             case 8:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.portraitAnimTimer = 0x18;
                 self->step = 6;
                 return;
             case 9:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue++;
                     g_Dialogue.nextCharDialogue++;
                     continue;
@@ -218,7 +273,7 @@ void EntityMariaCutscene(Entity* self) {
                 g_api.PlaySfx(nextChar);
                 continue;
             case 10:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != false) {
@@ -227,7 +282,7 @@ void EntityMariaCutscene(Entity* self) {
                 *g_Dialogue.nextCharDialogue--;
                 return;
             case 11:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 if (g_api.func_80131F68() != true) {
@@ -236,47 +291,48 @@ void EntityMariaCutscene(Entity* self) {
                 *g_Dialogue.nextCharDialogue--;
                 return;
             case 12:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                CutsceneUnk6((u8*)bit_shifty);
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                CutsceneUnk6((u8*)ptr);
                 continue;
             case 13:
                 continue;
             case 14:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty += (u16)0x100000;
-                g_Dialogue.nextCharDialogue += *(u16*)bit_shifty << 2;
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                // This needs help. Casting the const to short is good.
+                ptr += (u16)0x100000;
+                g_Dialogue.nextCharDialogue += *(u16*)ptr << 2;
 
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
 
-            case 15:
-                bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                bit_shifty <<= 4;
-                bit_shifty |= (s32)*g_Dialogue.nextCharDialogue;
-                g_Dialogue.nextCharDialogue = (u8*)bit_shifty + 0x100000;
+            case 15: // c64
+                ptr = (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue++;
+                ptr <<= 4;
+                ptr |= (u_long)*g_Dialogue.nextCharDialogue;
+                g_Dialogue.nextCharDialogue = (u8*)ptr + 0x100000;
                 continue;
 
             case 16:
@@ -294,20 +350,20 @@ void EntityMariaCutscene(Entity* self) {
                 g_Dialogue.unk3C = 0;
                 continue;
             case 19:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     g_Dialogue.nextCharDialogue += 5;
                 } else {
-                    bit_shifty = (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty <<= 4;
-                    bit_shifty |= (s32)*g_Dialogue.nextCharDialogue++;
-                    bit_shifty += 0x100000;
+                    ptr = (u_long)*g_Dialogue.nextCharDialogue++;  // 8e4
+                    ptr <<= 4;                                     // 918
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++; // 91c
+                    ptr <<= 4;                                     // 920
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++; // 924
+                    ptr <<= 4;                                     // 928
+                    ptr |= (u_long)*g_Dialogue.nextCharDialogue++; // 92c
+                    ptr += 0x100000;                               //
                     nextChar2 = g_Dialogue.nextCharDialogue++[0];
-                    LoadTPage((u32*)bit_shifty, 1, 0, D_801813D4[nextChar2],
-                              0x100, 0x30, 0x48);
+                    LoadTPage((u32*)ptr, 1, 0, D_80180790[nextChar2], 0x100,
+                              0x30, 0x48);
                 }
                 continue;
             case 20:
@@ -316,31 +372,17 @@ void EntityMariaCutscene(Entity* self) {
                 nextChar |= *g_Dialogue.nextCharDialogue++;
                 g_api.PlaySfx(nextChar);
                 continue;
-            case 21:
-                g_CutsceneFlags = 0;
-                D_801CB684 = 0;
-                D_801CB73C = 0;
-                continue;
-            case 22:
-                g_CutsceneFlags &= ~(1 << *g_Dialogue.nextCharDialogue++);
-                continue;
-            case 23:
-                return;
-            case 24:
-                if (!((g_CutsceneFlags >> *g_Dialogue.nextCharDialogue) & 1)) {
-                    *g_Dialogue.nextCharDialogue--;
-                    return;
-                }
-                *g_Dialogue.nextCharDialogue++;
-                continue;
+            // Note: Other cutscenes have Case 21-24 here, this one is missing
+            // it.
             default:
-                if (D_801CB684 != 0) {
+                if (skip_cutscene) {
                     continue;
                 }
                 g_Dialogue.nextCharTimer = g_Dialogue.unk17;
             }
             break; // This breaks the big while-loop!
         }
+        // End of while loop, comes at a58
         if (nextChar == 0x20) {
             g_Dialogue.nextCharX += 2;
             return;
@@ -366,6 +408,10 @@ void EntityMariaCutscene(Entity* self) {
         prim->x1 = prim->x3 += 4;
         prim->y0 = prim->y1 -= 6;
         prim->y2 = prim->y3 += 6;
+        // This block is in case 4 for other cutscenes, case 3 for this one
+        if (prim->x1 >= (g_Dialogue.startX - 2)) {
+            prim->x1 = prim->x3 = g_Dialogue.startX - 3;
+        }
         g_Dialogue.portraitAnimTimer--;
         if (!g_Dialogue.portraitAnimTimer) {
             self->step = 1;
@@ -381,9 +427,6 @@ void EntityMariaCutscene(Entity* self) {
         prim->x1 = prim->x3 -= 4;
         prim->y0 = prim->y1 += 6;
         prim->y2 = prim->y3 -= 6;
-        if (prim->x1 >= (g_Dialogue.startX - 2)) {
-            prim->x1 = prim->x3 = g_Dialogue.startX - 3;
-        }
         g_Dialogue.portraitAnimTimer--;
         if (!g_Dialogue.portraitAnimTimer) {
             self->step = 1;
@@ -393,7 +436,7 @@ void EntityMariaCutscene(Entity* self) {
     case 5:
         switch (self->step_s) {
         case 0:
-            D_801CB73C = 1;
+            D_801C2580 = 1;
             primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 0x48);
             if (primIndex == -1) {
                 DestroyEntity(self);
@@ -408,14 +451,14 @@ void EntityMariaCutscene(Entity* self) {
                 prim->y0 = prim->y1 = g_Dialogue.startY + j;
                 prim->priority = 0x1FE;
                 prim->drawMode = DRAW_DEFAULT;
-                prim->x2 = D_801813DC[j];
+                prim->x2 = D_80180798[j];
                 prim->x3 = 0xF70;
 
                 j++;
             }
             self->step_s++;
             return;
-        case 1:
+        case 1: /* switch 3 */
             for (j = 0, prim = &g_PrimBuf[g_Dialogue.primIndex[0]];
                  prim != NULL; prim = prim->next) {
                 prim->x3 -= prim->x2;
@@ -493,10 +536,9 @@ void EntityMariaCutscene(Entity* self) {
         break;
 
     case 7:
-        DestroyEntity(self);
-        g_CastleFlags[CASTLE_FLAG_133] = 1;
         D_8003C704 = 0;
-        g_Settings.D_8003CB04 |= 0x100;
+        g_Settings.D_8003CB04 |= 1;
+        DestroyEntity(self);
         break;
     }
 }
